Do not use short ints for station IDs in data table. Fixes #72

+++ dballe/db/v6/mysql/driver.cc
@@ -152,7 +152,7 @@ void Driver::create_tables_v6()
     conn.exec_no_data(R"(
         CREATE TABLE data (
            id          INTEGER auto_increment PRIMARY KEY,
-           id_station  SMALLINT NOT NULL,
+           id_station  INTEGER NOT NULL,
            id_report   INTEGER NOT NULL,
            id_lev_tr   INTEGER NOT NULL,
            datetime    DATETIME NOT NULL,
